Add 'eCSL' status updates on LCD.
It can indicate if door-lock has enh. CSL status 'disabled', 'enabled', 'linking', 'linked'.

Disabled: Door-lock is not sampling for the wake-up frames.
Enabled : Door-lock is sampling for the wake-up frames.
Linking : Door-lock is in the process of linking with Wake-up parent.
Linked  : Attached with the wake-up parent and ready to communicate over enh. CSL.